Fix bugs: sorting columns and display of action columns
Fix the bug which showed both multiple action columns and
single action column (action columns have action components
that perfom actions on data. It can show one column with
all action components, or one column for each action or none).

Fix the bug that prevented sorting the columns when clicking
on the column's title. The bug was fixed by refactoring the
props of DataTable.vue. Now, the props will be parsed only when
the columns or the text is updated.

Lastly, change styles of sorting icons in order to align them
with text.