fix bug when sorting numeric data
The compare function used to sort numbers was
returning true or false values depending on
if a number was greater than other. However, it
should return a negative number if the first number
is greater than the second and a positive number
otherwise. That caused the data to not be sort
properly. Fix that.