add language service provider
now the user can set the translation for the specified language,
even if there is no default translation. That way, users themselves
can provide custom translations that can be used globally accross
any instance of VueDataTanle

Note: even though the language can be accessed globally, we are NOT
using global variables, localStorage, sessionStorage or any stuff
like that. Instead, we are using functions to edit variables that are
scoped to VueDataTable. That way, the users cannot access those variables
directly; they can only interact with them through an interface. That
way, with the help of webpack, we isolate the variables to the scope
of this library only.