use component to render table cell by default
Before, the table cell content was rendered as text, or was rendered  by
a custom component provided by the user of the package. Now,  the  table
cell content is also rendered via a component, but a component from  the
package itself it the user did not provided a custom component.